tests/submission/PhilipTaylor/canvas/size.attributes.style.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.style</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.style</h1>
<p class="desc">Canvas size is independent of CSS resizing</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.css">size.css</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="30" id="c" style="width: 100px; height: 50px" width="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="size.attributes.style.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Canvas size is independent of CSS resizing");
_addTest(function(canvas, ctx) {

_assertEqual(canvas.width, 50, "canvas.width", "50");
_assertEqual(canvas.height, 30, "canvas.height", "30");


});
</script>

<div id="log"></div>
</body></html>
